[REFACTOR] 관심사 분리 #353
